requirements.txt
appdirs==1.4.4
attrs==20.1.0
black==19.10b0
click==7.1.2
flake8==3.8.3
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
pathspec==0.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
redis==3.5.3
regex==2020.7.14
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
